feat(file-drop): make file-drop optional with build tag `filedrop`
