+++ README.md
@@ -47,7 +47,7 @@ git clone --recurse-submodules https://github.com/KaijuEngine/kaiju.git
 If you have Go, C build tools, platform libs, and Vulkan setup, you can start by running:
 ```sh
 cd src
-go build -tags="debug,editor" -o ../ ./
+go build -tags="debug,editor,filedrop" -o ../ ./
 ```
 
 *Or just open the repository in VSCode (or other IDE) and begin debugging it.*
