+++ src/editor/editor_file_drop_router_stub.go
@@ -1,0 +1,13 @@
+//go:build !editor && !filedrop
+
+/******************************************************************************/
+/* editor_file_drop_router_stub.go                                            */
+/******************************************************************************/
+
+package editor
+
+type FileDropRouter struct{}
+
+func (ed *Editor) FileDropRouter() *FileDropRouter { return &ed.fileDropRouter }
+
+func (ed *Editor) connectFileDropRouter() {}
