+++ src/platform/windowing/window_filedrop_stub.go
@@ -1,0 +1,9 @@
+//go:build !editor && !filedrop
+
+/******************************************************************************/
+/* window_filedrop_stub.go                                                    */
+/******************************************************************************/
+
+package windowing
+
+type fileDropModule struct{}
